+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * added more type hints
 * `stubbles\db\config\DatabaseConfiguration::getInitialQuery()` now returns an empty string if no initial query configured
 * parameter `$type` of `stubbles\db\Statement::bindParam()` and `stubbles\db\Statement::bindValue()` must be an int
+* `stubbles\db\DatabaseConnection::getLastInsertId()` now returns a `string` instead of `int`
 
 ## 8.0.0 (2016-07-31)